mpi/topo: split MPIR implementations from MPI functions
