+++ src/include/mpir_topo.h
@@ -87,6 +87,7 @@ int MPIR_Cart_create_impl(MPIR_Comm * comm_ptr, int ndims, const int dims[],
                           const int periods[], int reorder, MPI_Comm * comm_cart);
 int MPIR_Cart_map_impl(const MPIR_Comm * comm_ptr, int ndims, const int dims[],
                        const int periodic[], int *newrank);
+int MPIR_Dims_create_impl(int nnodes, int ndims, int dims[]);
 
 int MPIR_Graph_create(MPIR_Comm *, int, const int[], const int[], int, MPI_Comm *);
 int MPIR_Graph_map(const MPIR_Comm *, int, const int[], const int[], int *);
@@ -100,5 +101,10 @@ int MPIR_Dist_graph_neighbors_count_impl(MPIR_Comm * comm_ptr, int *indegree, in
 int MPIR_Dist_graph_neighbors_impl(MPIR_Comm * comm_ptr, int maxindegree, int sources[],
                                    int sourceweights[], int maxoutdegree, int destinations[],
                                    int destweights[]);
+
+int MPIR_Cart_coords(MPIR_Comm * comm_ptr, int rank, int maxdims, int coords[]);
+int MPIR_Cart_get(MPIR_Comm * comm_ptr, int maxdims, int dims[], int periods[], int coords[]);
+int MPIR_Cart_sub(MPIR_Comm * comm_ptr, const int remain_dims[], MPI_Comm * newcomm);
+int MPIR_Graph_get(MPIR_Comm * comm_ptr, int maxindex, int maxedges, int indx[], int edges[]);
 
 #endif /* MPIR_TOPO_H_INCLUDED */
