+++ src/mpi/topo/cart_create.c
@@ -27,164 +27,6 @@ int MPI_Cart_create(MPI_Comm comm_old, int ndims, const int dims[], const int pe
 
 #endif
 
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-int MPIR_Cart_create(MPIR_Comm * comm_ptr, int ndims, const int dims[],
-                     const int periods[], int reorder, MPI_Comm * comm_cart)
-{
-    int i, newsize, rank, nranks, mpi_errno = MPI_SUCCESS;
-    MPIR_Comm *newcomm_ptr = NULL;
-    MPIR_Topology *cart_ptr = NULL;
-    MPIR_CHKPMEM_DECL(4);
-
-    /* Set this as null incase we exit with an error */
-    *comm_cart = MPI_COMM_NULL;
-
-    /* Check for invalid arguments */
-    newsize = 1;
-    for (i = 0; i < ndims; i++)
-        newsize *= dims[i];
-
-    /* Use ERR_ARG instead of ERR_TOPOLOGY because there is no topology yet */
-    MPIR_ERR_CHKANDJUMP2((newsize > comm_ptr->remote_size), mpi_errno,
-                         MPI_ERR_ARG, "**cartdim",
-                         "**cartdim %d %d", comm_ptr->remote_size, newsize);
-
-    if (ndims == 0) {
-        /* specified as a 0D Cartesian topology in MPI 2.1.
-         * Rank 0 returns a dup of COMM_SELF with the topology info attached.
-         * Others return MPI_COMM_NULL. */
-
-        rank = comm_ptr->rank;
-
-        if (rank == 0) {
-            MPIR_Comm *comm_self_ptr;
-            MPIR_Comm_get_ptr(MPI_COMM_SELF, comm_self_ptr);
-            mpi_errno = MPIR_Comm_dup_impl(comm_self_ptr, NULL, &newcomm_ptr);
-            MPIR_ERR_CHECK(mpi_errno);
-
-            /* Create the topology structure */
-            MPIR_CHKPMEM_MALLOC(cart_ptr, MPIR_Topology *, sizeof(MPIR_Topology),
-                                mpi_errno, "cart_ptr", MPL_MEM_COMM);
-
-            cart_ptr->kind = MPI_CART;
-            cart_ptr->topo.cart.nnodes = 1;
-            cart_ptr->topo.cart.ndims = 0;
-
-            /* make mallocs of size 1 int so that they get freed as part of the
-             * normal free mechanism */
-
-            MPIR_CHKPMEM_MALLOC(cart_ptr->topo.cart.dims, int *, sizeof(int),
-                                mpi_errno, "cart.dims", MPL_MEM_COMM);
-            MPIR_CHKPMEM_MALLOC(cart_ptr->topo.cart.periodic, int *, sizeof(int),
-                                mpi_errno, "cart.periodic", MPL_MEM_COMM);
-            MPIR_CHKPMEM_MALLOC(cart_ptr->topo.cart.position, int *, sizeof(int),
-                                mpi_errno, "cart.position", MPL_MEM_COMM);
-        } else {
-            *comm_cart = MPI_COMM_NULL;
-            goto fn_exit;
-        }
-    } else {
-
-        /* Create a new communicator as a duplicate of the input communicator
-         * (but do not duplicate the attributes) */
-        if (reorder) {
-
-            /* Allow the cart map routine to remap the assignment of ranks to
-             * processes */
-            mpi_errno = MPIR_Cart_map_impl(comm_ptr, ndims, (const int *) dims,
-                                           (const int *) periods, &rank);
-            MPIR_ERR_CHECK(mpi_errno);
-
-            /* Create the new communicator with split, since we need to reorder
-             * the ranks (including the related internals, such as the connection
-             * tables */
-            mpi_errno = MPIR_Comm_split_impl(comm_ptr,
-                                             rank == MPI_UNDEFINED ? MPI_UNDEFINED : 1,
-                                             rank, &newcomm_ptr);
-            MPIR_ERR_CHECK(mpi_errno);
-
-        } else {
-            mpi_errno = MPII_Comm_copy((MPIR_Comm *) comm_ptr, newsize, NULL, &newcomm_ptr);
-            MPIR_ERR_CHECK(mpi_errno);
-            rank = comm_ptr->rank;
-        }
-
-        /* If this process is not in the resulting communicator, return a
-         * null communicator and exit */
-        if (rank >= newsize || rank == MPI_UNDEFINED) {
-            *comm_cart = MPI_COMM_NULL;
-            goto fn_exit;
-        }
-
-        /* Create the topololgy structure */
-        MPIR_CHKPMEM_MALLOC(cart_ptr, MPIR_Topology *, sizeof(MPIR_Topology),
-                            mpi_errno, "cart_ptr", MPL_MEM_COMM);
-
-        cart_ptr->kind = MPI_CART;
-        cart_ptr->topo.cart.nnodes = newsize;
-        cart_ptr->topo.cart.ndims = ndims;
-        MPIR_CHKPMEM_MALLOC(cart_ptr->topo.cart.dims, int *, ndims * sizeof(int),
-                            mpi_errno, "cart.dims", MPL_MEM_COMM);
-        MPIR_CHKPMEM_MALLOC(cart_ptr->topo.cart.periodic, int *, ndims * sizeof(int),
-                            mpi_errno, "cart.periodic", MPL_MEM_COMM);
-        MPIR_CHKPMEM_MALLOC(cart_ptr->topo.cart.position, int *, ndims * sizeof(int),
-                            mpi_errno, "cart.position", MPL_MEM_COMM);
-        nranks = newsize;
-        for (i = 0; i < ndims; i++) {
-            cart_ptr->topo.cart.dims[i] = dims[i];
-            cart_ptr->topo.cart.periodic[i] = periods[i];
-            nranks = nranks / dims[i];
-            /* FIXME: nranks could be zero (?) */
-            cart_ptr->topo.cart.position[i] = rank / nranks;
-            rank = rank % nranks;
-        }
-    }
-
-
-    /* Place this topology onto the communicator */
-    mpi_errno = MPIR_Topology_put(newcomm_ptr, cart_ptr);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIR_OBJ_PUBLISH_HANDLE(*comm_cart, newcomm_ptr->handle);
-
-  fn_exit:
-    return mpi_errno;
-
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-    MPIR_CHKPMEM_REAP();
-    /* --END ERROR HANDLING-- */
-    goto fn_exit;
-}
-
-int MPIR_Cart_create_impl(MPIR_Comm * comm_ptr, int ndims, const int dims[],
-                          const int periods[], int reorder, MPI_Comm * comm_cart)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (comm_ptr->topo_fns != NULL && comm_ptr->topo_fns->cartCreate != NULL) {
-        /* --BEGIN USEREXTENSION-- */
-        mpi_errno = comm_ptr->topo_fns->cartCreate(comm_ptr, ndims,
-                                                   (const int *) dims,
-                                                   (const int *) periods, reorder, comm_cart);
-        MPIR_ERR_CHECK(mpi_errno);
-        /* --END USEREXTENSION-- */
-    } else {
-        mpi_errno = MPIR_Cart_create(comm_ptr, ndims,
-                                     (const int *) dims, (const int *) periods, reorder, comm_cart);
-        MPIR_ERR_CHECK(mpi_errno);
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-#endif
-
 /*@
 
 MPI_Cart_create - Makes a new communicator to which topology information
