+++ src/mpi/topo/cart_shift.c
@@ -23,61 +23,6 @@ int MPI_Cart_shift(MPI_Comm comm, int direction, int disp, int *rank_source, int
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Cart_shift
 #define MPI_Cart_shift PMPI_Cart_shift
-
-int MPIR_Cart_shift_impl(MPIR_Comm * comm_ptr, int direction, int disp, int *rank_source,
-                         int *rank_dest)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Topology *cart_ptr;
-    int i;
-    int pos[MAX_CART_DIM];
-
-    cart_ptr = MPIR_Topology_get(comm_ptr);
-
-    MPIR_ERR_CHKANDJUMP((!cart_ptr ||
-                         cart_ptr->kind != MPI_CART), mpi_errno, MPI_ERR_TOPOLOGY, "**notcarttopo");
-    MPIR_ERR_CHKANDJUMP((cart_ptr->topo.cart.ndims == 0), mpi_errno, MPI_ERR_TOPOLOGY,
-                        "**dimszero");
-    MPIR_ERR_CHKANDJUMP2((direction >= cart_ptr->topo.cart.ndims), mpi_errno, MPI_ERR_ARG,
-                         "**dimsmany", "**dimsmany %d %d", cart_ptr->topo.cart.ndims, direction);
-
-    /* Check for the case of a 0 displacement */
-    if (disp == 0) {
-        *rank_source = *rank_dest = comm_ptr->rank;
-    } else {
-        /* To support advanced implementations that support MPI_Cart_create,
-         * we compute the new position and call PMPI_Cart_rank to get the
-         * source and destination.  We could bypass that step if we know that
-         * the mapping is trivial.  Copy the current position. */
-        for (i = 0; i < cart_ptr->topo.cart.ndims; i++) {
-            pos[i] = cart_ptr->topo.cart.position[i];
-        }
-        /* We must return MPI_PROC_NULL if shifted over the edge of a
-         * non-periodic mesh */
-        pos[direction] += disp;
-        if (!cart_ptr->topo.cart.periodic[direction] &&
-            (pos[direction] >= cart_ptr->topo.cart.dims[direction] || pos[direction] < 0)) {
-            *rank_dest = MPI_PROC_NULL;
-        } else {
-            MPIR_Cart_rank_impl(cart_ptr, pos, rank_dest);
-        }
-
-        pos[direction] = cart_ptr->topo.cart.position[direction] - disp;
-        if (!cart_ptr->topo.cart.periodic[direction] &&
-            (pos[direction] >= cart_ptr->topo.cart.dims[direction] || pos[direction] < 0)) {
-            *rank_source = MPI_PROC_NULL;
-        } else {
-            MPIR_Cart_rank_impl(cart_ptr, pos, rank_source);
-        }
-    }
-
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
 
 /*@
