+++ src/mpi/topo/graphcreate.c
@@ -32,97 +32,6 @@ int MPI_Graph_create(MPI_Comm comm_old, int nnodes, const int indx[], const int 
    solicit warning messages from the compiler.  To avoid those, the name
    indx is used instead of index in this file */
 
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-
-int MPIR_Graph_create(MPIR_Comm * comm_ptr, int nnodes,
-                      const int indx[], const int edges[], int reorder, MPI_Comm * comm_graph)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int i, nedges;
-    MPIR_Comm *newcomm_ptr = NULL;
-    MPIR_Topology *graph_ptr = NULL;
-    MPIR_CHKPMEM_DECL(3);
-
-    /* Set this to null in case there is an error */
-    *comm_graph = MPI_COMM_NULL;
-
-    /* Create a new communicator */
-    if (reorder) {
-        int nrank;
-
-        /* Allow the cart map routine to remap the assignment of ranks to
-         * processes */
-        mpi_errno = MPIR_Graph_map_impl(comm_ptr, nnodes, indx, edges, &nrank);
-        MPIR_ERR_CHECK(mpi_errno);
-        /* Create the new communicator with split, since we need to reorder
-         * the ranks (including the related internals, such as the connection
-         * tables */
-        mpi_errno = MPIR_Comm_split_impl(comm_ptr,
-                                         nrank == MPI_UNDEFINED ? MPI_UNDEFINED : 1,
-                                         nrank, &newcomm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else {
-        /* Just use the first nnodes processes in the communicator */
-        mpi_errno = MPII_Comm_copy((MPIR_Comm *) comm_ptr, nnodes, NULL, &newcomm_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-    }
-
-
-    /* If this process is not in the resulting communicator, return a
-     * null communicator and exit */
-    if (!newcomm_ptr) {
-        *comm_graph = MPI_COMM_NULL;
-        goto fn_exit;
-    }
-
-    nedges = indx[nnodes - 1];
-    MPIR_CHKPMEM_MALLOC(graph_ptr, MPIR_Topology *, sizeof(MPIR_Topology),
-                        mpi_errno, "graph_ptr", MPL_MEM_COMM);
-
-    graph_ptr->kind = MPI_GRAPH;
-    graph_ptr->topo.graph.nnodes = nnodes;
-    graph_ptr->topo.graph.nedges = nedges;
-    MPIR_CHKPMEM_MALLOC(graph_ptr->topo.graph.index, int *,
-                        nnodes * sizeof(int), mpi_errno, "graph.index", MPL_MEM_COMM);
-    MPIR_CHKPMEM_MALLOC(graph_ptr->topo.graph.edges, int *,
-                        nedges * sizeof(int), mpi_errno, "graph.edges", MPL_MEM_COMM);
-    for (i = 0; i < nnodes; i++)
-        graph_ptr->topo.graph.index[i] = indx[i];
-    for (i = 0; i < nedges; i++)
-        graph_ptr->topo.graph.edges[i] = edges[i];
-
-    /* Finally, place the topology onto the new communicator and return the
-     * handle */
-    mpi_errno = MPIR_Topology_put(newcomm_ptr, graph_ptr);
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-
-    MPIR_OBJ_PUBLISH_HANDLE(*comm_graph, newcomm_ptr->handle);
-
-    /* ... end of body of routine ... */
-
-  fn_exit:
-    return mpi_errno;
-
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-    MPIR_CHKPMEM_REAP();
-#ifdef HAVE_ERROR_CHECKING
-    {
-        mpi_errno =
-            MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE, __func__, __LINE__, MPI_ERR_OTHER,
-                                 "**mpi_graph_create", "**mpi_graph_create %C %d %p %p %d %p",
-                                 comm_ptr->handle, nnodes, indx, edges, reorder, comm_graph);
-    }
-#endif
-    mpi_errno = MPIR_Err_return_comm((MPIR_Comm *) comm_ptr, __func__, mpi_errno);
-    goto fn_exit;
-    /* --END ERROR HANDLING-- */
-}
-#endif
-
 /*@
 MPI_Graph_create - Makes a new communicator to which topology information
                  has been attached
